Add appprotect_common package
