+++ internal/k8s/appprotect/app_protect_configuration.go
@@ -5,6 +5,8 @@ import (
 	"sort"
 	"time"
 
+	"github.com/nginxinc/kubernetes-ingress/internal/k8s/appprotect_common"
+
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 	"k8s.io/apimachinery/pkg/runtime/schema"
 )
@@ -341,7 +343,7 @@ func (ci *ConfigurationImpl) verifyPolicyAgainstUserSigs(policy *PolicyEx) bool 
 
 // AddOrUpdatePolicy adds or updates an App Protect Policy to App Protect Configuration
 func (ci *ConfigurationImpl) AddOrUpdatePolicy(policyObj *unstructured.Unstructured) (changes []Change, problems []Problem) {
-	resNsName := GetNsName(policyObj)
+	resNsName := appprotect_common.GetNsName(policyObj)
 	policy, err := createAppProtectPolicyEx(policyObj)
 	if err != nil {
 		ci.Policies[resNsName] = policy
@@ -361,7 +363,7 @@ func (ci *ConfigurationImpl) AddOrUpdatePolicy(policyObj *unstructured.Unstructu
 
 // AddOrUpdateLogConf adds or updates App Protect Log Configuration to App Protect Configuration
 func (ci *ConfigurationImpl) AddOrUpdateLogConf(logconfObj *unstructured.Unstructured) (changes []Change, problems []Problem) {
-	resNsName := GetNsName(logconfObj)
+	resNsName := appprotect_common.GetNsName(logconfObj)
 	logConf, err := createAppProtectLogConfEx(logconfObj)
 	ci.LogConfs[resNsName] = logConf
 	if err != nil {
@@ -373,7 +375,7 @@ func (ci *ConfigurationImpl) AddOrUpdateLogConf(logconfObj *unstructured.Unstruc
 
 // AddOrUpdateUserSig adds or updates App Protect User Defined Signature to App Protect Configuration
 func (ci *ConfigurationImpl) AddOrUpdateUserSig(userSigObj *unstructured.Unstructured) (change UserSigChange, problems []Problem) {
-	resNsName := GetNsName(userSigObj)
+	resNsName := appprotect_common.GetNsName(userSigObj)
 	userSig, err := createAppProtectUserSigEx(userSigObj)
 	ci.UserSigs[resNsName] = userSig
 	if err != nil {
@@ -563,7 +565,7 @@ func NewFakeConfiguration() Configuration {
 
 // AddOrUpdatePolicy adds or updates an App Protect Policy to App Protect Configuration
 func (fc *FakeConfiguration) AddOrUpdatePolicy(policyObj *unstructured.Unstructured) (changes []Change, problems []Problem) {
-	resNsName := GetNsName(policyObj)
+	resNsName := appprotect_common.GetNsName(policyObj)
 	policy := &PolicyEx{
 		Obj:     policyObj,
 		IsValid: true,
@@ -574,7 +576,7 @@ func (fc *FakeConfiguration) AddOrUpdatePolicy(policyObj *unstructured.Unstructu
 
 // AddOrUpdateLogConf adds or updates App Protect Log Configuration to App Protect Configuration
 func (fc *FakeConfiguration) AddOrUpdateLogConf(logConfObj *unstructured.Unstructured) (changes []Change, problems []Problem) {
-	resNsName := GetNsName(logConfObj)
+	resNsName := appprotect_common.GetNsName(logConfObj)
 	logConf := &LogConfEx{
 		Obj:     logConfObj,
 		IsValid: true,
