+++ internal/k8s/appprotectdos/app_protect_dos_resources.go
@@ -6,6 +6,8 @@ import (
 	"regexp"
 	"strconv"
 	"strings"
+	
+	"github.com/nginxinc/kubernetes-ingress/internal/k8s/appprotect_common"
 
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
@@ -19,60 +21,12 @@ var appProtectDosLogConfRequiredFields = [][]string{
 	{"spec", "filter"},
 }
 
-func validateRequiredFields(policy *unstructured.Unstructured, fieldsList [][]string) error {
-	for _, fields := range fieldsList {
-		field, found, err := unstructured.NestedMap(policy.Object, fields...)
-		if err != nil {
-			return fmt.Errorf("Error checking for required field %v: %w", field, err)
-		}
-		if !found {
-			return fmt.Errorf("Required field %v not found", field)
-		}
-	}
-	return nil
-}
-
 // ValidateAppDosProtectLogConf validates LogConfiguration resource
-func ValidateAppProtectDosLogConf(logConf *unstructured.Unstructured) error {
+func validateAppProtectDosLogConf(logConf *unstructured.Unstructured) error {
 	lcName := logConf.GetName()
-	err := validateRequiredFields(logConf, appProtectDosLogConfRequiredFields)
+	err := appprotect_common.ValidateRequiredFields(logConf, appProtectDosLogConfRequiredFields)
 	if err != nil {
 		return fmt.Errorf("Error validating App Protect Dos Log Configuration %v: %w", lcName, err)
-	}
-
-	return nil
-}
-
-var (
-	logDstEx     = regexp.MustCompile(`(?:syslog:server=((?:\d{1,3}\.){3}\d{1,3}|localhost):\d{1,5})|stderr`)
-)
-
-// ValidateAppProtectDosLogDestination validates destination for log configuration
-func ValidateAppProtectDosLogDestination(dstAntn string) error {
-	errormsg := "Error parsing App Protect Dos Log config: Destination must follow format: syslog:server=<ip-address | localhost>:<port> or stderr"
-	if !logDstEx.MatchString(dstAntn) {
-		return fmt.Errorf("%s Log Destination did not follow format", errormsg)
-	}
-	if dstAntn == "stderr" {
-		return nil
-	}
-
-	dstchunks := strings.Split(dstAntn, ":")
-
-	// This error can be ignored since the regex check ensures this string will be parsable
-	port, _ := strconv.Atoi(dstchunks[2])
-
-	if port > 65535 || port < 1 {
-		return fmt.Errorf("Error parsing port: %v not a valid port number", port)
-	}
-
-	ipstr := strings.Split(dstchunks[1], "=")[1]
-	if ipstr == "localhost" {
-		return nil
-	}
-
-	if net.ParseIP(ipstr) == nil {
-		return fmt.Errorf("Error parsing host: %v is not a valid ip address", ipstr)
 	}
 
 	return nil
@@ -107,23 +61,10 @@ func ValidateAppProtectDosAccessLogDest(accessLogDest string) error {
 func validateAppProtectDosPolicy(policy *unstructured.Unstructured) error {
 	polName := policy.GetName()
 
-	err := validateRequiredFields(policy, appProtectDosPolicyRequiredFields)
+	err := appprotect_common.ValidateRequiredFields(policy, appProtectDosPolicyRequiredFields)
 	if err != nil {
 		return fmt.Errorf("Error validating App Protect Dos Policy %v: %w", polName, err)
 	}
 
 	return nil
-}
-
-// ParseResourceReferenceAnnotation returns a namespace/name string
-func ParseResourceReferenceAnnotation(ns, antn string) string {
-	if !strings.Contains(antn, "/") {
-		return ns + "/" + antn
-	}
-	return antn
-}
-
-// GetNsName gets the key of a resource in the format: "resNamespace/resName"
-func GetNsName(obj *unstructured.Unstructured) string {
-	return obj.GetNamespace() + "/" + obj.GetName()
 }
