+++ internal/k8s/appprotectdos/app_protect_dos_resources_test.go
@@ -7,35 +7,6 @@ import (
 	"k8s.io/apimachinery/pkg/apis/meta/v1/unstructured"
 )
 
-func TestValidateAppProtectDsLogDestinationAnnotation(t *testing.T) {
-	// Positive test cases
-	posDstAntns := []string{"stderr", "syslog:server=localhost:9000", "syslog:server=10.1.1.2:9000", "/var/log/ap.log"}
-
-	// Negative test cases item, expected error message
-	negDstAntns := [][]string{
-		{"stdout", "Log Destination did not follow format"},
-		{"syslog:server=localhost:99999", "not a valid port number"},
-		{"syslog:server=999.99.99.99:5678", "is not a valid ip address"},
-	}
-
-	for _, tCase := range posDstAntns {
-		err := ValidateAppProtectDosLogDestination(tCase)
-		if err != nil {
-			t.Errorf("got %v expected nil", err)
-		}
-	}
-	for _, nTCase := range negDstAntns {
-		err := ValidateAppProtectDosLogDestination(nTCase[0])
-		if err == nil {
-			t.Errorf("got no error expected error containing %s", nTCase[1])
-		} else {
-			if !strings.Contains(err.Error(), nTCase[1]) {
-				t.Errorf("got %v expected to contain: %s", err, nTCase[1])
-			}
-		}
-	}
-}
-
 func TestValidateAppProtectDosAccessLogDest(t *testing.T) {
 	// Positive test cases
 	posDstAntns := []string{"10.10.1.1:514"}
@@ -121,7 +92,7 @@ func TestValidateAppProtectDosLogConf(t *testing.T) {
 	}
 
 	for _, test := range tests {
-		err := ValidateAppProtectDosLogConf(test.logConf)
+		err := validateAppProtectDosLogConf(test.logConf)
 		if test.expectErr && err == nil {
 			t.Errorf("validateAppProtectDosLogConf() returned no error for the case of %s", test.msg)
 		}
